backend/requirements.txt
# Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.7.0

# Database
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.0

# AI & ML
openai==1.58.0
sentence-transformers==3.3.1
numpy>=1.26.0

# File Processing
PyPDF2==3.0.1
python-docx==1.1.2
Pillow==11.0.0
pytesseract==0.3.13

# File Watching
watchdog==6.0.0

# Utilities
python-multipart==0.0.19
websockets==14.1
httpx==0.28.1
sse-starlette==2.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
